+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Updated `tiny-loop` dependency from 0.4 to 0.5
+
 ## [0.4.0] - 2026-02-10
 
 ### Fixed